chore: Change file structure of StatsPage children, wrapped on a new PricesStats component
