+++ app/components/PricesStats/PricesStats.tsx
@@ -1,0 +1,17 @@
+import { PricesAverages } from './PricesAverages/PricesAverages';
+import { PricesTable } from './PricesTable/PricesTable';
+
+import { CountryPrice } from '@/lib/types';
+
+interface PricesStatsProps {
+  data: CountryPrice[];
+}
+
+export function PricesStats({ data }: PricesStatsProps) {
+  return (
+    <>
+      <PricesAverages data={data} />
+      <PricesTable data={data} />
+    </>
+  );
+}
